add startup sound to prime audio element
There's a noticeable lag on first play that seems to be related to the file size of the first item to be played. I'm trying out priming the audio element by immediately playing a small startup sound on initial load. Probably even better would be a really tiny subsecond blank mp3. For now it's the sound of a needle dropping on a 45 RPM LP.