Properly restore now-playing when window is closed
Closes https://github.com/hypermodules/hyperamp/issues/196

It does introduce a bug though.  Because there is only one trackOrder
array, when we change the track order array, the metadata also get
updated, showing the wrong song info for whats playing.  It corrects on
the following queue, or if you restore whatever tracklist you had
before.

Fundamentally, we need to fix the data structure behind play queus and
the view state. We need to persist the active play state order, and only
activate a new play order state when a song is queued from a different
track order.

This issue is described more in https://github.com/hypermodules/hyperamp/issues/191